requirements.txt
bitcoin-requests
requests
psycopg2
